requirements.txt
mutagen
hurry.filesize
pyusb
psutil
prompt-toolkit
